backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
mutagen==1.47.0
watchdog==3.0.0
pillow==10.1.0
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
